Invoice/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djoser>=2.2
django-filter>=23.2
django-cors-headers>=4.3
